backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai==0.8.3
Pillow==10.2.0
PyMuPDF==1.23.8
scikit-learn==1.4.0
numpy==1.26.3
python-dotenv==1.0.0
requests==2.31.0
